created examples for to demo opengl, spi and audio (fft)
