+++ src/MCP3008Comm.cpp
@@ -1,3 +1,11 @@
+/** @file MCP3008Comm.cpp
+ * 
+ * @brief TO WRITE
+ *
+ * @author Marcell Illyes (marcellillyes), Davide Rovelli (daviderovell0)
+ * 
+ */
+
 #include "MCP3008Comm.h"
