--- src/control/Makefile
@@ -1,238 +1,0 @@
-# CMAKE generated file: DO NOT EDIT!
-# Generated by "Unix Makefiles" Generator, CMake Version 3.10
-
-# Default target executed when no arguments are given to make.
-default_target: all
-
-.PHONY : default_target
-
-# Allow only one "make -f Makefile2" at a time, but pass parallelism.
-.NOTPARALLEL:
-
-
-#=============================================================================
-# Special targets provided by cmake.
-
-# Disable implicit rules so canonical targets will work.
-.SUFFIXES:
-
-
-# Remove some rules from gmake that .SUFFIXES does not remove.
-SUFFIXES =
-
-.SUFFIXES: .hpux_make_needs_suffix_list
-
-
-# Suppress display of executed commands.
-$(VERBOSE).SILENT:
-
-
-# A target that is always out of date.
-cmake_force:
-
-.PHONY : cmake_force
-
-#=============================================================================
-# Set environment variables for the build.
-
-# The shell in which to execute make rules.
-SHELL = /bin/sh
-
-# The CMake executable.
-CMAKE_COMMAND = /usr/bin/cmake
-
-# The command to remove a file.
-RM = /usr/bin/cmake -E remove -f
-
-# Escaping for special characters.
-EQUALS = =
-
-# The top-level source directory on which CMake was run.
-CMAKE_SOURCE_DIR = /home/davide/repositories/bzzzbz/control
-
-# The top-level build directory on which CMake was run.
-CMAKE_BINARY_DIR = /home/davide/repositories/bzzzbz/control
-
-#=============================================================================
-# Targets provided globally by CMake.
-
-# Special rule for the target rebuild_cache
-rebuild_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "Running CMake to regenerate build system..."
-	/usr/bin/cmake -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR)
-.PHONY : rebuild_cache
-
-# Special rule for the target rebuild_cache
-rebuild_cache/fast: rebuild_cache
-
-.PHONY : rebuild_cache/fast
-
-# Special rule for the target edit_cache
-edit_cache:
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --cyan "No interactive CMake dialog available..."
-	/usr/bin/cmake -E echo No\ interactive\ CMake\ dialog\ available.
-.PHONY : edit_cache
-
-# Special rule for the target edit_cache
-edit_cache/fast: edit_cache
-
-.PHONY : edit_cache/fast
-
-# The main all target
-all: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/davide/repositories/bzzzbz/control/CMakeFiles /home/davide/repositories/bzzzbz/control/CMakeFiles/progress.marks
-	$(MAKE) -f CMakeFiles/Makefile2 all
-	$(CMAKE_COMMAND) -E cmake_progress_start /home/davide/repositories/bzzzbz/control/CMakeFiles 0
-.PHONY : all
-
-# The main clean target
-clean:
-	$(MAKE) -f CMakeFiles/Makefile2 clean
-.PHONY : clean
-
-# The main clean target
-clean/fast: clean
-
-.PHONY : clean/fast
-
-# Prepare targets for installation.
-preinstall: all
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall
-
-# Prepare targets for installation.
-preinstall/fast:
-	$(MAKE) -f CMakeFiles/Makefile2 preinstall
-.PHONY : preinstall/fast
-
-# clear depends
-depend:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 1
-.PHONY : depend
-
-#=============================================================================
-# Target rules for targets named spi_test
-
-# Build rule for target.
-spi_test: cmake_check_build_system
-	$(MAKE) -f CMakeFiles/Makefile2 spi_test
-.PHONY : spi_test
-
-# fast build rule for target.
-spi_test/fast:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/build
-.PHONY : spi_test/fast
-
-MCP3008Comm.o: MCP3008Comm.cpp.o
-
-.PHONY : MCP3008Comm.o
-
-# target to build an object file
-MCP3008Comm.cpp.o:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/MCP3008Comm.cpp.o
-.PHONY : MCP3008Comm.cpp.o
-
-MCP3008Comm.i: MCP3008Comm.cpp.i
-
-.PHONY : MCP3008Comm.i
-
-# target to preprocess a source file
-MCP3008Comm.cpp.i:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/MCP3008Comm.cpp.i
-.PHONY : MCP3008Comm.cpp.i
-
-MCP3008Comm.s: MCP3008Comm.cpp.s
-
-.PHONY : MCP3008Comm.s
-
-# target to generate assembly for a file
-MCP3008Comm.cpp.s:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/MCP3008Comm.cpp.s
-.PHONY : MCP3008Comm.cpp.s
-
-gpio-sysfs.o: gpio-sysfs.cpp.o
-
-.PHONY : gpio-sysfs.o
-
-# target to build an object file
-gpio-sysfs.cpp.o:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/gpio-sysfs.cpp.o
-.PHONY : gpio-sysfs.cpp.o
-
-gpio-sysfs.i: gpio-sysfs.cpp.i
-
-.PHONY : gpio-sysfs.i
-
-# target to preprocess a source file
-gpio-sysfs.cpp.i:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/gpio-sysfs.cpp.i
-.PHONY : gpio-sysfs.cpp.i
-
-gpio-sysfs.s: gpio-sysfs.cpp.s
-
-.PHONY : gpio-sysfs.s
-
-# target to generate assembly for a file
-gpio-sysfs.cpp.s:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/gpio-sysfs.cpp.s
-.PHONY : gpio-sysfs.cpp.s
-
-spi_test.o: spi_test.cpp.o
-
-.PHONY : spi_test.o
-
-# target to build an object file
-spi_test.cpp.o:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/spi_test.cpp.o
-.PHONY : spi_test.cpp.o
-
-spi_test.i: spi_test.cpp.i
-
-.PHONY : spi_test.i
-
-# target to preprocess a source file
-spi_test.cpp.i:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/spi_test.cpp.i
-.PHONY : spi_test.cpp.i
-
-spi_test.s: spi_test.cpp.s
-
-.PHONY : spi_test.s
-
-# target to generate assembly for a file
-spi_test.cpp.s:
-	$(MAKE) -f CMakeFiles/spi_test.dir/build.make CMakeFiles/spi_test.dir/spi_test.cpp.s
-.PHONY : spi_test.cpp.s
-
-# Help Target
-help:
-	@echo "The following are some of the valid targets for this Makefile:"
-	@echo "... all (the default if no target is provided)"
-	@echo "... clean"
-	@echo "... depend"
-	@echo "... rebuild_cache"
-	@echo "... spi_test"
-	@echo "... edit_cache"
-	@echo "... MCP3008Comm.o"
-	@echo "... MCP3008Comm.i"
-	@echo "... MCP3008Comm.s"
-	@echo "... gpio-sysfs.o"
-	@echo "... gpio-sysfs.i"
-	@echo "... gpio-sysfs.s"
-	@echo "... spi_test.o"
-	@echo "... spi_test.i"
-	@echo "... spi_test.s"
-.PHONY : help
-
-
-
-#=============================================================================
-# Special targets to cleanup operation of make.
-
-# Special rule to run CMake to check the build system integrity.
-# No rule that depends on this can have commands that come from listfiles
-# because they might be regenerated.
-cmake_check_build_system:
-	$(CMAKE_COMMAND) -H$(CMAKE_SOURCE_DIR) -B$(CMAKE_BINARY_DIR) --check-build-system CMakeFiles/Makefile.cmake 0
-.PHONY : cmake_check_build_system
-
