+++ src/examples/Makefile
