+++ src/shaders/spectrum_slow.glsl
@@ -1,3 +1,11 @@
+/** @file spectrum_slow.glsl
+ * 
+ * @brief An earlier implementation of spectrum.glsl, that caused performance issues, demonstrates the limitations of loops in GLSL.
+ *
+ * @author Marcell Illyes (marcellillyes)
+ * 
+ */
+
 uniform float H;
 uniform float W;
 uniform float fft[513];
@@ -11,23 +19,15 @@ void main( void ) {
 	float color = 0.0;
 	float c= 0.0;
 	
-	
 	for(int i=0; i<=512;i+=16){
 	
 		if(mod(X,1.0/32.0)<0.02 && X<c/32.0 && X > (c-1.0)/32.0){
-		
 			if(Y<fft[i]){
-			
-				color+=0.5;
-				
-			}
-			
+				color+=0.5;	
+			}	
 		}
-		
 		c+=1.0;
 	}
-	
-
 	gl_FragColor = vec4( vec3( color, color,color ), 1.0 );
 
 }
