+++ src/shaders/vertex.glsl
@@ -1,3 +1,11 @@
+/** @file vertex.glsl
+ * 
+ * @brief Vertex shader. Fixed for BZZZBZ.
+ *
+ * @author Peter Nagy (deetrone)
+ * 
+ */
+
 attribute vec3 coord3d;
 
 void main(void) {
