+++ src/shaders/wavepatterns.glsl
@@ -1,29 +1,31 @@
-uniform float H;
-uniform float W;
-uniform float c_A;
-const float pi = 3.141596;
+/** @file wavepatterns.glsl
+ * 
+ * @brief Example shader to demonstrate SPI control of BZZZBZ.
+ *
+ * @author Marcell Illyes (marcellillyes)
+ * 
+ */
 
-float BPM = 120.0;
-float THICKK = 0.0015;
-float RES = 100.0;
-float AMP = 0.5;
-float FREQ = 5.0;
-float BIAS = 0.5 ;
-float PHASE = 0.2 * pi;
+
+uniform float W;
+uniform float H;
+uniform float A;
+uniform float B;
+uniform float C;
+uniform float time;
+
+const float pi = 3.1415926;
+const float THICK = 0.0015;
+const float RES = 100.0;
+const float FREQ = 5.0;
 
 
 void main( void ) {
-   
-    float X = gl_FragCoord.x/W;
-    float Y = gl_FragCoord.y/H;
-	float color = 0.0;
-	
-	if(mod(AMP * cos(Y * FREQ * pi - PHASE) + BIAS + AMP * sin(X * FREQ * pi - PHASE) + BIAS, 1.0 / RES) <= THICKK){
-	
-	        color += 1.0+c_A;
-		
-	}
-	
-	gl_FragColor = vec4(color, (0.6*Y)*color, 0.2*color, 1.0 );
-
+    	float X = gl_FragCoord.x/W;
+    	float Y = gl_FragCoord.y/H;
+	    float color = 0.0;
+	    if(mod((1.5*A) * cos((0.2*time/24.0)+Y * (FREQ+B*0.2) * pi - 0.2*pi) + (1.5*A) * sin((0.2*time/16.0)+X * (FREQ+B*0.3) * pi - 0.2*pi) , 1.0 / RES) <= THICK*(4.9*C-0.5)){
+	        color += 1.0;
+		}
+	    gl_FragColor = vec4(1.0*color, 0.5*Y*color+(0.2*B*Y*color), 0.1*color, 1.0 );
 }
