--- src/video_processing/Backup/window.v.glsl
@@ -1,5 +1,0 @@
-attribute vec3 coord3d;
-
-void main(void) {
-  gl_Position = vec4(coord3d, 1.0);
-}
