--- src/video_processing/old/colortest/LOpenGL.h
@@ -1,13 +1,0 @@
-/*This source code copyrighted by Lazy Foo' Productions (2004-2013)
-and may not be redistributed without written permission.*/
-//Version: 001
-
-#ifndef LOPENGL_H
-#define LOPENGL_H
-
-#include <GL/freeglut.h>
-#include <GL/gl.h>
-#include <GL/glu.h>
-#include <stdio.h>
-
-#endif
