--- src/video_processing/old/main.cpp
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-int main(){
-
-printf("Hello Video");
-
-}
