--- src/video_processing/old/triangle disp/LUtil.h
@@ -1,53 +1,0 @@
-/*This source code copyrighted by Lazy Foo' Productions (2004-2013)
-and may not be redistributed without written permission.*/
-//Version: 001
-
-#ifndef LUTIL_H
-#define LUTIL_H
-
-#include "LOpenGL.h"
-#include <stdio.h>
-
-//Screen Constants
-const int SCREEN_WIDTH = 640;
-const int SCREEN_HEIGHT = 480;
-const int SCREEN_FPS = 60;
-
-bool initGL();
-/*
-Pre Condition:
- -A valid OpenGL context
-Post Condition:
- -Initializes matrices and clear color
- -Reports to console if there was an OpenGL error
- -Returns false if there was an error in initialization
-Side Effects:
- -Projection matrix is set to identity matrix
- -Modelview matrix is set to identity matrix
- -Matrix mode is set to modelview
- -Clear color is set to black
-*/
-
-void update();
-/*
-Pre Condition:
- -None
-Post Condition:
- -Does per frame logic
-Side Effects:
- -None
-*/
-
-void render();
-/*
-Pre Condition:
- -A valid OpenGL context
- -Active modelview matrix
-Post Condition:
- -Renders the scene
-Side Effects:
- -Clears the color buffer
- -Swaps the front/back buffer
-*/
-
-#endif
