--- src/video_processing/old/triangle disp/readme.txt
@@ -1,24 +1,0 @@
-g++ LUtil.cpp main.cpp -w -lGL -lGLU -lglut -o 01_hello_freeglut
-
-Libraries required for OpenGL:
-
-sudo apt-get install freeglut3
-sudo apt-get install freeglut3-dev
-
-
-Some stuff from original author.
-
-Copyright Notice:
------------------
-The files within this zip file are copyrighted by Lazy Foo' Productions (2004-2013)
-and may not be redistributed without written permission.
-
-This project is linked against:
-----------------------------------------
-Windows:
-OpenGL32
-freeglut
-
-*nix:
-GL
-freeglut
