--- src/video_processing/shaders/vertex.glsl
@@ -1,7 +1,0 @@
-//simple vertex shader, only static coordinates, no transformations etc
-
-attribute vec3 coord3d;
-
-void main(void) {
-  gl_Position = vec4(coord3d, 1.0);
-}
